requirements.txt
Package             Version
------------------- -------
asgiref             3.5.2  
Django              3.2    
django-filter       21.1   
djangorestframework 3.13.1 
importlib-metadata  4.11.3 
Markdown            3.3.7  
pip                 20.0.2 
pkg-resources       0.0.0  
pytz                2022.1 
setuptools          44.0.0 
sqlparse            0.4.2  
wheel               0.34.2 
zipp                3.8.0  
